+++ assignment/bai1/Officer.java
@@ -2,11 +2,11 @@ package assignment.bai1;
 
 public class Officer {
   protected String name;
-  protected short age;
+  protected int age;
   protected String gender;
   protected String address;
 
-  public Officer(String name, short age, String gender, String address) {
+  public Officer(String name, int age, String gender, String address) {
     this.name = name;
     this.age = age;
     this.gender = gender;
